refactor: Error 이름 변경
